backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
drf-spectacular
psycopg2-binary
python-decouple
gunicorn
pdfplumber
pytesseract
pdf2image
celery
redis
